requirements.txt
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.1
pip==9.0.1
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
setuptools==40.8.0
sqlparse==0.3.0

